fixing type error

+++ makani/models/networks/fourcastnet3.py
@@ -815,7 +815,7 @@ class AtmoSphericNeuralOperatorNet(nn.Module):
             # the following eventually leads to spectral instability
             # w = nn.functional.softplus(x[..., self.water_channels, :, :], beta=5, threshold=5)
             w_full = torch.zeros_like(x)
-            w_full.index_copy_(-3, self.water_channels, w)
+            w_full.index_copy_(-3, self.water_channels, w.to(x.dtype))
             x = torch.where(self.water_channel_mask, w_full, x)
 
         return x
